+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.1.0-alpha.2] - 2021-7-27
+- add option dragWidgetBuilder, can config custom drag widget.
+- add option scrollSpeedController, can config the scroll speed.
+
 ## [1.1.0-alpha.1] - 2021-7-26
 - fix childAspectRatio bug